fixing price

+++ ChargingApp/DTOs/NewOrderDto.cs
@@ -7,7 +7,7 @@ public class NewOrderDto
     [Required] public string? PlayerId { get; set; }
     [Required] public int ProductId { get; set; }
     [Required] public string PlayerName { get; set; }
-    public double Quantity { get; set; } = 1;
+    public decimal Quantity { get; set; } = 1;
 }
 
 public class NewNormalOrderDto : NewOrderDto
