+++ ChargingApp/DTOs/OrderAdminDto.cs
@@ -6,9 +6,9 @@ public class OrderAdminDto
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
     public string ArabicName { get; set; }
     public string EnglishName { get; set; }
-    public double TotalPrice { get; set; }
+    public decimal TotalPrice { get; set; }
     public string? PlayerId { get; set; }
-    public double Quantity { get; set; } = 1;
+    public decimal Quantity { get; set; } = 1;
     public string OrderType { get; set; } = "Normal";
     public string Notes { get; set; }= "Pending";
 }
