+++ ChargingApp/Entity/AppUser.cs
@@ -10,11 +10,11 @@ public class AppUser : IdentityUser<int>
     public string? LastName { get; set; }
     public string? Country { get; set; }
     public string? City { get; set; }
-    public double Balance { get; set; } = 0;
+    public decimal Balance { get; set; } = 0;
     public int VIPLevel { get; set; } = 0;
-    public double TotalPurchasing { get; set; } = 0;
-    public double Debit { get; set; } = 0;
-    public double TotalForVIPLevel { get; set; } = 0;
+    public decimal TotalPurchasing { get; set; } = 0;
+    public decimal Debit { get; set; } = 0;
+    public decimal TotalForVIPLevel { get; set; } = 0;
     public ICollection<AppUserRole> UserRoles { get; set; }
 
 }
