+++ ChargingApp/Entity/BenefitPercentInSpecificVilLevel.cs
@@ -4,5 +4,5 @@ public class BenefitPercentInSpecificVilLevel : BaseEntity
 {
     public int ProductId { get; set; }
     public int VipLevel { get; set; }
-    public double BenefitPercent { get; set; }
+    public decimal BenefitPercent { get; set; }
 }
