+++ ChargingApp/Entity/DebitHistory.cs
@@ -2,7 +2,7 @@
 
 public class DebitHistory : BaseEntity
 {
-    public double DebitValue { get; set; }
+    public decimal DebitValue { get; set; }
     public AppUser User { get; set; }
     public DateTime Date { get; set; } = DateTime.Now;
 }
