+++ ChargingApp/Entity/Product.cs
@@ -4,8 +4,8 @@ public class Product :BaseEntity
 {
     public string? EnglishName { get; set; }
     public string? ArabicName { get; set; }
-    public double Price { get; set; }
-    public double Quantity { get; set; } = 0;
+    public decimal Price { get; set; }
+    public decimal Quantity { get; set; } = 0;
     public bool CanChooseQuantity { get; set; } = false;
     public bool Available { get; set; } = true;
     public Photo? Photo { get; set; }
