+++ ChargingApp/Interfaces/IUserRepository.cs
@@ -7,5 +7,5 @@ public interface IUserRepository
     public Task<AppUser?> GetUserByEmailAsync(string? email);
     public void UpdateUserInfo(AppUser user) ;
     public Task<bool> DeleteUserByEmail(string? email);
-    public Task<double> GetBenefitPercentAsync(string? email);
+    public Task<decimal> GetBenefitPercentAsync(string? email);
 }
